+++ BUILD.gn
@@ -1310,7 +1310,6 @@ config("grpc_config") {
         "include/grpcpp/resource_quota.h",
         "include/grpcpp/security/auth_context.h",
         "include/grpcpp/security/auth_metadata_processor.h",
-        "include/grpcpp/security/auth_metadata_processor_impl.h",
         "include/grpcpp/security/credentials.h",
         "include/grpcpp/security/credentials_impl.h",
         "include/grpcpp/security/server_credentials.h",
