+++ CMakeLists.txt
@@ -2891,7 +2891,6 @@ foreach(_hdr
   include/grpcpp/resource_quota.h
   include/grpcpp/security/auth_context.h
   include/grpcpp/security/auth_metadata_processor.h
-  include/grpcpp/security/auth_metadata_processor_impl.h
   include/grpcpp/security/credentials.h
   include/grpcpp/security/credentials_impl.h
   include/grpcpp/security/server_credentials.h
@@ -3588,7 +3587,6 @@ foreach(_hdr
   include/grpcpp/resource_quota.h
   include/grpcpp/security/auth_context.h
   include/grpcpp/security/auth_metadata_processor.h
-  include/grpcpp/security/auth_metadata_processor_impl.h
   include/grpcpp/security/credentials.h
   include/grpcpp/security/credentials_impl.h
   include/grpcpp/security/server_credentials.h
