+++ Makefile
@@ -5093,7 +5093,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/resource_quota.h \
     include/grpcpp/security/auth_context.h \
     include/grpcpp/security/auth_metadata_processor.h \
-    include/grpcpp/security/auth_metadata_processor_impl.h \
     include/grpcpp/security/credentials.h \
     include/grpcpp/security/credentials_impl.h \
     include/grpcpp/security/server_credentials.h \
@@ -5788,7 +5787,6 @@ PUBLIC_HEADERS_CXX += \
     include/grpcpp/resource_quota.h \
     include/grpcpp/security/auth_context.h \
     include/grpcpp/security/auth_metadata_processor.h \
-    include/grpcpp/security/auth_metadata_processor_impl.h \
     include/grpcpp/security/credentials.h \
     include/grpcpp/security/credentials_impl.h \
     include/grpcpp/security/server_credentials.h \
