+++ build_autogenerated.yaml
@@ -2418,7 +2418,6 @@ libs:
   - include/grpcpp/resource_quota.h
   - include/grpcpp/security/auth_context.h
   - include/grpcpp/security/auth_metadata_processor.h
-  - include/grpcpp/security/auth_metadata_processor_impl.h
   - include/grpcpp/security/credentials.h
   - include/grpcpp/security/credentials_impl.h
   - include/grpcpp/security/server_credentials.h
@@ -2807,7 +2806,6 @@ libs:
   - include/grpcpp/resource_quota.h
   - include/grpcpp/security/auth_context.h
   - include/grpcpp/security/auth_metadata_processor.h
-  - include/grpcpp/security/auth_metadata_processor_impl.h
   - include/grpcpp/security/credentials.h
   - include/grpcpp/security/credentials_impl.h
   - include/grpcpp/security/server_credentials.h
