+++ tools/doxygen/Doxyfile.c++.internal
@@ -1024,7 +1024,6 @@ include/grpcpp/impl/service_type.h \
 include/grpcpp/resource_quota.h \
 include/grpcpp/security/auth_context.h \
 include/grpcpp/security/auth_metadata_processor.h \
-include/grpcpp/security/auth_metadata_processor_impl.h \
 include/grpcpp/security/credentials.h \
 include/grpcpp/security/credentials_impl.h \
 include/grpcpp/security/server_credentials.h \
